# docs/api/openapi.yaml
components:
  schemas:
...
    AgeGroupingStrategy:
      description: Enumeration for age grouping strategies.
      enum:
      - yearly
      - custom
      title: AgeGroupingStrategy
      type: string
...
    AnalysisMethod:
      description: Enumeration for anomaly detection methods.
      enum:
      - autoencoder
      title: AnalysisMethod
      type: string
...
    ConfigurationUpdateRequest:
      description: Request model for updating system configuration.
      properties:
        age_grouping_strategy:
          anyOf:
          - $ref: '#/components/schemas/AgeGroupingStrategy'
          - type: 'null'
        max_age_group_span:
          anyOf:
          - exclusiveMinimum: 0.0
            maximum: 16.0
            type: number
          - type: 'null'
          title: Max Age Group Span
        min_samples_per_group:
          anyOf:
          - minimum: 10.0
            type: integer
          - type: 'null'
          title: Min Samples Per Group
        threshold_percentile:
          anyOf:
          - maximum: 99.9
            minimum: 50.0
            type: number
          - type: 'null'
          description: Percentile for threshold calculation
          title: Threshold Percentile
      title: ConfigurationUpdateRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    SuccessResponse:
      description: Standard success response model.
      properties:
        data:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Response data
          title: Data
        message:
          description: Success message
          title: Message
          type: string
        success:
          default: true
          title: Success
          type: boolean
      required:
      - message
      title: SuccessResponse
      type: object
    SystemConfigurationResponse:
      description: Response model for system configuration.
      properties:
        age_grouping_strategy:
          $ref: '#/components/schemas/AgeGroupingStrategy'
        anomaly_detection_method:
          $ref: '#/components/schemas/AnalysisMethod'
        max_age_group_span:
          exclusiveMinimum: 0.0
          maximum: 16.0
          title: Max Age Group Span
          type: number
        min_samples_per_group:
          minimum: 10.0
          title: Min Samples Per Group
          type: integer
        threshold_percentile:
          maximum: 99.9
          minimum: 50.0
          title: Threshold Percentile
          type: number
        vision_model:
          $ref: '#/components/schemas/VisionModel'
      required:
      - vision_model
      - anomaly_detection_method
      - threshold_percentile
      - age_grouping_strategy
      - min_samples_per_group
      - max_age_group_span
      title: SystemConfigurationResponse
      type: object
    ThresholdUpdateRequest:
      description: Request model for updating model thresholds.
      properties:
        percentile:
          anyOf:
          - maximum: 99.9
            minimum: 50.0
            type: number
          - type: 'null'
          description: Percentile for automatic threshold calculation
          title: Percentile
        threshold:
          description: New threshold value
          exclusiveMinimum: 0.0
          title: Threshold
          type: number
      required:
      - threshold
      title: ThresholdUpdateRequest
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    VisionModel:
      description: Enumeration for vision models.
      enum:
      - vit
      title: VisionModel
      type: string
info:
  contact:
    email: support@example.com
    name: API Support
  description: Machine learning system for detecting anomalies in children's drawings
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Children's Drawing Anomaly Detection System
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/config/:
    get:
      description: 'Get current system configuration.


        This endpoint returns the current system configuration including

        model settings, threshold parameters, and age grouping strategy.'
      operationId: get_config_api_v1_config__get
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/SystemConfigurationResponse'
          description: Successful Response
      summary: Get Config
      tags:
      - configuration
    put:
      description: 'Update system configuration.


        This endpoint updates various system configuration settings

        including thresholds and age grouping parameters.'
      operationId: update_config_api_v1_config__put
      requestBody:
        content:
          application/json:
            example: {}
            schema:
              $ref: '#/components/schemas/ConfigurationUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Config
      tags:
      - configuration
...
  /api/v1/config/reset:
    post:
      description: 'Reset system configuration and models.


        WARNING: This endpoint deactivates all models and clears caches.

        Use with caution in production environments.'
      operationId: reset_system_api_v1_config_reset_post
      parameters:
      - in: query
        name: confirm
        required: false
        schema:
          default: false
          title: Confirm
          type: boolean
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Reset System
      tags:
      - configuration
...
  /api/v1/config/threshold:
    put:
      description: 'Update global threshold settings.


        This endpoint recalculates thresholds for all active models

        using the specified percentile value from the request body.'
      operationId: update_threshold_settings_api_v1_config_threshold_put
      requestBody:
        content:
          application/json:
            example: {}
            schema:
              $ref: '#/components/schemas/ThresholdUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Threshold Settings
      tags:
      - configuration
...
servers:
- description: Development server
  url: http://localhost:8000
- description: API v1 base path
  url: /api/v1